moved the upper banner to signup and signin
